+++ bin/service-registry-cli.js
@@ -8,24 +8,24 @@ program
   .version(version, '-v, --version')
   .name('service-registry-cli')
   .command('pull')
-  .option('-h, --hosts <hosts>', '(Default: "127.0.0.1:2379") etcd host', '127.0.0.1:2379')
+  .option('-h, --host <host>', '(Default: "127.0.0.1:2379") etcd host', '127.0.0.1:2379')
   .option('-e, --env <environment>', '(Default: "dev") environment(dev, staging or prod) to register service.', 'dev')
   .action(async (opts) => {
     await tasks.pull({
       env: opts.env,
-      hosts: opts.hosts
+      host: opts.host
     })
   });
 
   program
   .command('register')
-  .option('-h, --hosts <hosts>', '(Default: "127.0.0.1:2379") etcd host', '127.0.0.1:2379')
+  .option('-h, --host <host>', '(Default: "127.0.0.1:2379") etcd host', '127.0.0.1:2379')
   .option('-e, --env <environment>', '(Default: "dev") environment(dev, staging or prod) to register service.', 'dev')
   .requiredOption('--service.endpoint <string>', 'service endpoint to register service.')
   .action(async (opts) => {
      await tasks.register({
       env: opts.env,
-      hosts: opts.hosts,
+      host: opts.host,
       serviceEndpoint: opts['service.endpoint']
      })
   });