+++ src/tasks/register.js
@@ -1,16 +1,18 @@
 const { default: ServiceRegistryFactory } = require('@hai.dinh/service-registry');
 const readJsonAtRootProject = require('../utils/readJsonFileAtRootProject');
 
-const register = async ({ env, hosts, serviceEndpoint }) => {
+const register = async ({ env, host, serviceEndpoint }) => {
   console.log(ServiceRegistryFactory, 'wtf');
   const registry = ServiceRegistryFactory.create({
     driver: 'etcd',
-    hosts,
+    host,
     env,
   });
 
   const serviceDefinition = await readJsonAtRootProject('service-definition.json');
   const packageInfo = await readJsonAtRootProject('package.json');
+
+  console.log(`Registering service with id: "${serviceDefinition.id}" ...`)
 
   await registry.register({
     id: serviceDefinition.id,
@@ -19,6 +21,8 @@ const register = async ({ env, hosts, serviceEndpoint }) => {
     endpoint: serviceEndpoint,
     version: packageInfo.version,
   })
+  
+  console.log('Done.')
 }
 
 module.exports = register;